# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vorea Studio API",
    "version": "2026-03-19",
    "summary": "API backend completa de Vorea Studio (inventario generado automáticamente).",
    "description": "Especificación OpenAPI generada a partir de rutas reales del backend. PayPal es la pasarela operativa actual."
  },
  "servers": [
    {
      "url": "http://localhost:5173",
      "description": "Frontend local"
    },
    {
      "url": "http://localhost:3001",
      "description": "API server local"
    }
  ],
...
  "paths": {
...
    "/api/paypal/client-id": {
      "get": {
        "operationId": "get_api_paypal_client_id",
        "tags": [
          "PayPal Orders"
        ],
        "summary": "Obtener /api/paypal/client-id",
        "description": "Endpoint real detectado en `server/app.ts:2681`.\n\nPolítica de acceso: **public**.\n\nDependencias primarias: paypal.\n\nAdvertencia: existe definición duplicada para este método/path en código.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          }
        },
        "x-vorea-meta": {
          "authPolicy": "public",
          "role": "public",
          "status": "duplicated_definition",
          "dependencies": [
            "paypal"
          ],
          "source": "server/app.ts:2681"
        }
      }
    },
...
    "/api/subscriptions/client-id": {
      "get": {
        "operationId": "get_api_subscriptions_client_id",
        "tags": [
          "PayPal Subscriptions"
        ],
        "summary": "Obtener /api/subscriptions/client-id",
        "description": "Endpoint real detectado en `server/paypal-subscriptions.ts:137`.\n\nPolítica de acceso: **public**.\n\nDependencias primarias: paypal.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          }
        },
        "x-vorea-meta": {
          "authPolicy": "public",
          "role": "public",
          "status": "active",
          "dependencies": [
            "paypal"
          ],
          "source": "server/paypal-subscriptions.ts:137"
        }
      }
    },
    "/api/subscriptions/create": {
      "post": {
        "operationId": "post_api_subscriptions_create",
        "tags": [
          "PayPal Subscriptions"
        ],
        "summary": "Crear/ejecutar /api/subscriptions/create",
        "description": "Endpoint real detectado en `server/paypal-subscriptions.ts:142`.\n\nPolítica de acceso: **authenticated**.\n\nDependencias primarias: paypal, prisma.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          },
          "403": {
            "description": "Sin permisos suficientes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          }
        },
        "x-vorea-meta": {
          "authPolicy": "authenticated",
          "role": "authenticated_user",
          "status": "active",
          "dependencies": [
            "paypal",
            "prisma"
          ],
          "source": "server/paypal-subscriptions.ts:142"
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              },
              "example": {
                "tier": "pro",
                "billing": "monthly"
              }
            }
          }
        }
      }
    },
    "/api/subscriptions/my-subscription": {
      "get": {
        "operationId": "get_api_subscriptions_my_subscription",
        "tags": [
          "PayPal Subscriptions"
        ],
        "summary": "Obtener /api/subscriptions/my-subscription",
        "description": "Endpoint real detectado en `server/paypal-subscriptions.ts:261`.\n\nPolítica de acceso: **public**.\n\nDependencias primarias: paypal, prisma.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          }
        },
        "x-vorea-meta": {
          "authPolicy": "public",
          "role": "public",
          "status": "active",
          "dependencies": [
            "paypal",
            "prisma"
          ],
          "source": "server/paypal-subscriptions.ts:261"
        }
      }
    },
    "/api/subscriptions/plans": {
      "get": {
        "operationId": "get_api_subscriptions_plans",
        "tags": [
          "PayPal Subscriptions"
        ],
        "summary": "Listar /api/subscriptions/plans",
        "description": "Endpoint real detectado en `server/paypal-subscriptions.ts:132`.\n\nPolítica de acceso: **public**.\n\nDependencias primarias: paypal.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          }
        },
        "x-vorea-meta": {
          "authPolicy": "public",
          "role": "public",
          "status": "active",
          "dependencies": [
            "paypal"
          ],
          "source": "server/paypal-subscriptions.ts:132"
        }
      }
    },
    "/api/subscriptions/webhook": {
      "post": {
        "operationId": "post_api_subscriptions_webhook",
        "tags": [
          "PayPal Subscriptions"
        ],
        "summary": "Crear/ejecutar /api/subscriptions/webhook",
        "description": "Endpoint real detectado en `server/paypal-subscriptions.ts:286`.\n\nPolítica de acceso: **public**.\n\nDependencias primarias: kv, paypal, prisma.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          }
        },
        "x-vorea-meta": {
          "authPolicy": "public",
          "role": "public",
          "status": "active",
          "dependencies": [
            "kv",
            "paypal",
            "prisma"
          ],
          "source": "server/paypal-subscriptions.ts:286"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              },
              "example": {
                "note": "Consultar contrato detallado en matriz operativa."
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "GenericSuccess": {
        "type": "object",
        "additionalProperties": true,
        "description": "Respuesta JSON exitosa (forma varía por endpoint)."
      },
      "GenericError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": true,
        "description": "Respuesta JSON de error."
      }
    }
  },
...
}